sharing data across classes
